prediction-api/requirements.txt
Flask==3.0.0
redis==5.0.1
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
python-dateutil==2.8.2
Werkzeug==3.0.1
gunicorn==21.2.0
kafka-python==2.0.2
